Add assimp_interface to the list of exported targets
The assimp_interface was implemented and added to the list of exported targets. This has been done to provide a cleaner and more encapsulated way of managing the assimp dependencies inside the VirtualRobot project. Now, instead of directly linking the assimp libraries to the VirtualRobot target, an interface library (assimp_interface) has been created to manage the assimp dependencies. This library is then linked to the VirtualRobot target. Additionally, install targets have been updated to include the assimp_interface.